Move EntityDialog to Entity/Dialog and some minor formatting changes
