+++ ShaRPG/Entity/Components/DialogComponent.cs
@@ -1,11 +1,10 @@
 ﻿using ShaRPG.Entity.Components.Messages;
-using ShaRPG.EntityDialog;
 
 namespace ShaRPG.Entity.Components {
     public class DialogComponent : AbstractComponent, IMessageHandler<MouseClickMessage> {
-        private readonly Dialog _dialog;
+        private readonly Dialog.Dialog _dialog;
         
-        public DialogComponent(GameEntity entity, Dialog dialog) : base(entity) {
+        public DialogComponent(GameEntity entity, Dialog.Dialog dialog) : base(entity) {
             _dialog = dialog;
         }
