+++ ShaRPG/Entity/Dialog/Dialog.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Xml.Linq;
 using System.Xml.XPath;
 
-namespace ShaRPG.EntityDialog {
+namespace ShaRPG.Entity.Dialog {
     public class Dialog {
         public readonly string Name;
         public string Graphic => $"ui_avatar_{_graphicName}";
@@ -17,10 +17,10 @@ namespace ShaRPG.EntityDialog {
         private int _currentNode = 0;
 
         public Dialog(string name, string graphic, IOpenDialog dialogOpener, Dictionary<int, DialogNode> dialogNodes) {
-            Name = name;
-            _graphicName = graphic;
+            Name          = name;
+            _graphicName  = graphic;
             _dialogOpener = dialogOpener;
-            _dialogNodes = dialogNodes;
+            _dialogNodes  = dialogNodes;
         }
 
         public void StartDialog() {
@@ -29,11 +29,11 @@ namespace ShaRPG.EntityDialog {
 
         public static Dialog FromXElement(string name, string graphic, XElement dialog, IOpenDialog dialogOpener) {
             Dictionary<int, DialogReply> replies = new Dictionary<int, DialogReply>();
-            Dictionary<int, DialogNode> nodes = new Dictionary<int, DialogNode>();
+            Dictionary<int, DialogNode>  nodes   = new Dictionary<int, DialogNode>();
 
             foreach (XElement node in dialog.XPathSelectElements("./Nodes/Node")) {
-                int nodeId = int.Parse(node.Attribute("id")?.Value);
-                string prompt = node.Elements("Prompt").FirstOrDefault()?.Value.Trim() ?? "";
+                int               nodeId      = int.Parse(node.Attribute("id")?.Value);
+                string            prompt      = node.Elements("Prompt").FirstOrDefault()?.Value.Trim() ?? "";
                 List<DialogReply> nodeReplies = new List<DialogReply>();
 
                 foreach (XElement nodeReply in node.Elements("Reply")) {
@@ -45,20 +45,20 @@ namespace ShaRPG.EntityDialog {
                         XElement reply = dialog.XPathSelectElement($"./Replies/Reply[@id='{replyId}']");
                         replies.Add(replyId, LoadReply(reply));
                     }
-                    
+
                     nodeReplies.Add(replies[replyId]);
                 }
-                
+
                 nodes.Add(nodeId, new DialogNode(nodeReplies, prompt));
             }
-            
+
             return new Dialog(name, graphic, dialogOpener, nodes);
         }
 
         private static DialogReply LoadReply(XElement reply) {
-            string prompt = reply.Elements("Prompt").FirstOrDefault()?.Value.Trim() ?? "";
+            string             prompt       = reply.Elements("Prompt").FirstOrDefault()?.Value.Trim() ?? "";
             List<DialogAction> replyActions = new List<DialogAction>();
-            
+
             foreach (XElement replyAction in reply.Elements("Action")) {
                 switch (replyAction.Attribute("type")?.Value) {
                     case "changeNode":
@@ -73,7 +73,7 @@ namespace ShaRPG.EntityDialog {
                         break;
                 }
             }
-            
+
             return new DialogReply(prompt, replyActions);
         }
 
@@ -88,7 +88,7 @@ namespace ShaRPG.EntityDialog {
         public void EndDialog() {
             OnEnd?.Invoke();
             _currentNode = 0;
-            OnEnd = null;
+            OnEnd        = null;
         }
     }
