+++ ShaRPG/Entity/Dialog/IOpenDialog.cs
@@ -1,4 +1,4 @@
-﻿namespace ShaRPG.EntityDialog {
+﻿namespace ShaRPG.Entity.Dialog {
     public interface IOpenDialog {
         void StartDialog(Dialog dialog);
     }
